# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/softwarepkg": {
            "get": {
                "description": "list software packages",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "list software packages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "phase of the softwarepkg",
                        "name": "phase",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name of the softwarepkg",
                        "name": "pkg_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "importer of the softwarePkg",
                        "name": "importer",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "platform of the softwarePkg",
                        "name": "platform",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "last software pkg id of previous page",
                        "name": "last_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "whether count total num of the pkgs",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page num which starts from 1",
                        "name": "page_num",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "count per page",
                        "name": "count_per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.SoftwarePkgSummariesDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            },
            "post": {
                "description": "apply a new software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "apply a new software package",
                "parameters": [
                    {
                        "description": "body of applying a new software package",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.softwarePkgRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app.NewSoftwarePkgDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/softwarepkg/:id": {
            "put": {
                "description": "update application of software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "update application of software package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of software package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body of updating software package application",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.reqToUpdateSoftwarePkg"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/softwarepkg/committers": {
            "post": {
                "description": "check committer of software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "check committer of software package",
                "parameters": [
                    {
                        "description": "body of checking committers",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.softwarePkgRepoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controller.checkCommittersResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/softwarepkg/{id}": {
            "get": {
                "description": "get software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "get software package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of software package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "language",
                        "name": "language",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.SoftwarePkgDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/softwarepkg/{id}/close": {
            "put": {
                "description": "close software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "close software package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of software package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "comment",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.reqToClosePkg"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/softwarepkg/{id}/retest": {
            "put": {
                "description": "rerun ci of software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "rerun ci of software package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of software package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/softwarepkg/{id}/review": {
            "get": {
                "description": "get user review on software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "get user review on software package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of software package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "language",
                        "name": "language",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.CheckItemUserReviewDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            },
            "post": {
                "description": "review software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "review software package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of software package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body of reviewing a software package",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.reviewRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.CheckItemReviewDTO": {
            "type": "object",
            "properties": {
                "agree": {
                    "type": "integer"
                },
                "desc": {
                    "type": "string"
                },
                "disagree": {
                    "type": "integer"
                },
                "has_result": {
                    "description": "HasResult if true, the pass is the result, otherwise the pass is meaningless.",
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "pass": {
                    "type": "boolean"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.UserCheckItemReviewDTO"
                    }
                }
            }
        },
        "app.CheckItemUserReviewDTO": {
            "type": "object",
            "properties": {
                "can_review": {
                    "type": "boolean"
                },
                "comment": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "pass": {
                    "type": "boolean"
                },
                "reviewed": {
                    "description": "Reviewed if true, the user reviewed this item before, otherwise the pass is meaningless.",
                    "type": "boolean"
                }
            }
        },
        "app.NewSoftwarePkgDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "app.SoftwarePkgCodeFileDTO": {
            "type": "object",
            "properties": {
                "download_addr": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "src": {
                    "type": "string"
                }
            }
        },
        "app.SoftwarePkgDTO": {
            "type": "object",
            "properties": {
                "applied_at": {
                    "type": "string"
                },
                "ci_status": {
                    "type": "string"
                },
                "committers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "importer": {
                    "type": "string"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.SoftwarePkgOperationLogDTO"
                    }
                },
                "phase": {
                    "type": "string"
                },
                "pkg_name": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "repo_link": {
                    "type": "string"
                },
                "reviews": {
                    "$ref": "#/definitions/app.SoftwarePkgReviewDTO"
                },
                "sig": {
                    "type": "string"
                },
                "spec": {
                    "$ref": "#/definitions/app.SoftwarePkgCodeFileDTO"
                },
                "srpm": {
                    "$ref": "#/definitions/app.SoftwarePkgCodeFileDTO"
                },
                "upstream": {
                    "type": "string"
                }
            }
        },
        "app.SoftwarePkgOperationLogDTO": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
        "app.SoftwarePkgReviewDTO": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.CheckItemReviewDTO"
                    }
                },
                "reviewers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "app.SoftwarePkgSummariesDTO": {
            "type": "object",
            "properties": {
                "pkgs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.SoftwarePkgSummaryDTO"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "app.SoftwarePkgSummaryDTO": {
            "type": "object",
            "properties": {
                "applied_at": {
                    "type": "string"
                },
                "ci_status": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "importer": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                },
                "pkg_name": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "repo_link": {
                    "type": "string"
                },
                "sig": {
                    "type": "string"
                }
            }
        },
...
        "controller.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "controller.checkCommittersResp": {
            "type": "object",
            "properties": {
                "invalid_committers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "controller.checkItemReviewInfo": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "pass": {
                    "type": "boolean"
                }
            }
        },
...
        "controller.reqToClosePkg": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                }
            }
        },
        "controller.reqToUpdateSoftwarePkg": {
            "type": "object",
            "properties": {
                "committers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "desc": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "repo_link": {
                    "type": "string"
                },
                "sig": {
                    "type": "string"
                },
                "spec_url": {
                    "type": "string"
                },
                "src_rpm_url": {
                    "type": "string"
                },
                "upstream": {
                    "type": "string"
                }
            }
        },
...
        "controller.reviewRequest": {
            "type": "object",
            "required": [
                "reviews"
            ],
            "properties": {
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.checkItemReviewInfo"
                    }
                }
            }
        },
        "controller.softwarePkgRepoRequest": {
            "type": "object",
            "required": [
                "repo_link"
            ],
            "properties": {
                "committers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "repo_link": {
                    "type": "string"
                }
            }
        },
        "controller.softwarePkgRequest": {
            "type": "object",
            "required": [
                "desc",
                "pkg_name",
                "reason",
                "repo_link",
                "sig",
                "spec_url",
                "src_rpm_url",
                "upstream"
            ],
            "properties": {
                "committers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "desc": {
                    "type": "string"
                },
                "pkg_name": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "repo_link": {
                    "type": "string"
                },
                "sig": {
                    "type": "string"
                },
                "spec_url": {
                    "type": "string"
                },
                "src_rpm_url": {
                    "type": "string"
                },
                "upstream": {
                    "type": "string"
                }
            }
        },
...
    }
}